change user type location
