--- nutri-track-client/src/views/Registration/types.ts
@@ -1,9 +1,0 @@
-export interface User {
-  email: string;
-  name: string;
-  password: string;
-  gender: string;
-  fitLevel: string;
-  weight: number;
-  height: number;
-}
